move correct sections + spacing fix
