+++ apps/roam/src/components/settings/GeneralSettings.tsx
@@ -8,27 +8,12 @@ import Description from "roamjs-components/components/Description";
 import { DEFAULT_CANVAS_PAGE_FORMAT } from "~/index";
 import refreshConfigTree from "~/utils/refreshConfigTree";
 
-const CANVAS_PAGE_FORMAT_KEY = "canvas-page-format";
-const DiscourseGraphHome = ({
-  extensionAPI,
-}: {
-  extensionAPI: OnloadArgs[`extensionAPI`];
-}) => {
+const DiscourseGraphHome = () => {
   const settings = useMemo(() => {
     refreshConfigTree();
     return getFormattedConfigTree();
   }, []);
-  const getInitCanvasPage = () => {
-    const savedFormat = extensionAPI.settings.get(
-      CANVAS_PAGE_FORMAT_KEY,
-    ) as string;
-    return savedFormat || DEFAULT_CANVAS_PAGE_FORMAT;
-  };
-  const [canvasPage, setCanvasPage] = useState(getInitCanvasPage);
-  const handleSetCanvasPage = (e: string) => {
-    extensionAPI.settings.set(CANVAS_PAGE_FORMAT_KEY, e);
-    setCanvasPage(e);
-  };
+
   return (
     <div className="flex flex-col gap-4 p-1">
       <div>
@@ -67,14 +52,6 @@ const DiscourseGraphHome = ({
         parentUid={settings.settingsUid}
         value={settings.trigger.value || ""}
       />
-      <Label>
-        Canvas Page Format
-        <Description description={"The page format for canvas pages"} />
-        <InputGroup
-          value={canvasPage}
-          onChange={(e) => handleSetCanvasPage(e.target.value)}
-        />
-      </Label>
     </div>
   );
 };
